fix http 400 status codes
After importing the geosparql files, we found that some URIs were returning HTTP 400 status codes. This was because the URIs had not been declared as Concepts, ConceptSchemes or Collections.

So the checkurlstatus.py python script checks which URIs are returning errors.
Then the createconcepts.py script creates skos:Concept instances for each URI.

So the createconcepts.py python script